+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+* `futures-enum` works well even if the dependency contains only sub-crates such as `futures-core`, `futures-util`, etc.
+
 # 0.1.2 - 2019-02-05
 
 * Update minimum derive_utils version to 0.6.3