+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+# 0.1.4 - 2019-02-21
+
 * Update to new nightly.
 
 # 0.1.3 - 2019-02-13